Add email verification for EmailAuth recovery requests
When the user submits an account recovery request, first send them a
confirmation email with a token. Only submit the recovery request to
Zendesk after they click the confirmation email.

The form data is stashed in MicroStash for 24 hours, but we require that
the user click the link within 15 minutes. If the link is expired but the
data is still in the stash, we resend the confirmation email with a new
token.

Bug: T407565
Change-Id: Ibf3817867d0c0965f8209c87b581144a6033b39b