Fetch user object from primary DB (for writes) not replica DB
When we're about to confirm the user's email during the secondary
authentication step, we need to ensure that the user object we're
operating with came from the primary DB. If not, we reload it from
the primary DB.

This is because we might perform DB writes in the GET request when
we're updating user information in the DB when the settings are
being saved.

If the user is not found in the primary DB, the code fails hard
because shouldn't have users in the replica and not in the primary.
`Special:ConfirmEmail` uses a similar approach.

Bug: T410652
Change-Id: I49fcec68427b70f30debab43adaa7f832afc66d5
(cherry picked from commit f674fc54418b9b2de5ae7698d338de4dfad1a556)